fix: prevent init cli side effects on core import
