+++ patterns/idp/index.ts
@@ -1,5 +1,5 @@
 export * from './feature-toggles/index.js';
 export * from './policy-engine/index.js';
 export * from './scorecards/index.js';
-export { initProject } from './init/cli.js';
+export { initProject } from './init/project.js';
 export * from './migration/index.js';
